re-org after consolidation
