--- client/package.json
@@ -1,73 +1,0 @@
-{
-  "name": "chatterbox",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@reduxjs/toolkit": "^1.8.1",
-    "autoprefixer": "^10.4.14",
-    "bootstrap": "^5.3.1",
-    "classnames": "^2.3.1",
-    "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.26",
-    "postcss-cli": "^10.1.0",
-    "postcss-nested": "^6.0.1",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-redux": "^8.0.1",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^5.0.1",
-    "redux-persist": "^6.0.0",
-    "socket.io-client": "^4.4.0",
-    "uuid": "^8.3.2"
-  },
-  "scripts": {
-    "build:css": "postcss src/styles/main.css -o src/index.css",
-    "watch:css": "postcss src/styles/main.css -o src/index.css -w",
-    "watch-node": "nodemon --watch src/server/listeners2.ts server",
-    "start": "react-scripts start",
-    "start:dev": "env-cmd -f ../.env.development react-scripts start",
-    "build-js": "BUILD_PATH='./dist' react-scripts build",
-    "build": "npm-run-all build:css build-js",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.3%",
-      "not ie 11",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version",
-      ">0.3%",
-      "not ie 11",
-      "not dead",
-      "not op_mini all"
-    ]
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.3",
-    "@types/node": "^20.4.8",
-    "@types/react": "^18.2.18",
-    "@types/react-dom": "^18.2.7",
-    "@types/react-redux": "^7.1.25",
-    "@types/react-router": "^5.1.20",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/redux-persist": "^4.3.1",
-    "@types/uuid": "^9.0.2",
-    "env-cmd": "^10.1.0",
-    "nodemon": "^3.0.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.1.6"
-  }
-}
