+++ package.json
@@ -31,8 +31,8 @@
     "watch:css": "postcss src/styles/main.css -o src/index.css -w",
     "watch-node": "nodemon --watch src/server/listeners2.ts server",
     "start": "react-scripts start",
-    "serve": "ts-node server/server.ts",
-    "dev": "env-cmd -f ../.env.development react-scripts start",
+    "serve": "ts-node src/server/server.ts",
+    "dev": "env-cmd -f .env.development react-scripts start",
     "build-js": "BUILD_PATH='./dist' react-scripts build",
     "build": "npm-run-all build:css build-js",
     "test": "react-scripts test --env=jsdom",
@@ -62,7 +62,7 @@
     ]
   },
   "devDependencies": {
-        "@types/express": "^4.17.17",
+    "@types/express": "^4.17.17",
     "@types/jest": "^29.5.3",
     "@types/node": "^20.4.8",
     "@types/react": "^18.2.18",
@@ -74,7 +74,7 @@
     "@types/uuid": "^9.0.2",
     "env-cmd": "^10.1.0",
     "nodemon": "^3.0.1",
-        "npm-run-all": "^4.1.5",
+    "npm-run-all": "^4.1.5",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.6"
   }
