--- server/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "chatterbox",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.17.2",
-    "socket.io": "^4.4.0",
-    "sqlite": "^5.1.1",
-    "sqlite3": "^5.1.6"
-  },
-  "scripts": {
-    "start": "ts-node server.ts",
-    "start:dev": "env-cmd -f ../.env.development ts-node server.ts"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.17",
-    "@types/node": "^20.4.8",
-    "env-cmd": "^10.1.0",
-    "nodemon": "^3.0.1",
-    "npm-run-all": "^4.1.5",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.1.6"
-  }
-}
